mysite/requirements.txt
asgiref==3.5.2
cart==1.2.1
contourpy==1.0.5
coverage==6.5.0
crispy==0.6.3
cycler==0.11.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.2
django-crispy-forms==1.14.0
django-import-export==2.9.0
django-tables2==2.4.1
et-xmlfile==1.1.0
fabio==0.14.0
fonttools==4.37.4
h5py==3.7.0
kiwisolver==1.4.4
MarkupPy==1.14
matplotlib==3.6.1
numpy==1.23.3
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
Pillow==9.2.0
pycryptodome==3.15.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
python-dateutil==2.8.2
PyYAML==6.0
silx==1.0.0
six==1.16.0
sqlparse==0.4.3
tablib==3.2.1
tzdata==2022.4
xlrd==2.0.1
xlwt==1.3.0
